+++ droit/plugins/output/eval.rand/main.py
@@ -3,6 +3,6 @@ import random
 # Choose a random answer
 # Example DDS v0.4: SRTX!danke->EVAL!rand.text(gerne, bitte)
 
-def text(data, db):
+def text(data, variables, db):
     rint = random.randint(0, len(data) - 1)
-    return data[rint], db
+    return data[rint], variables, db