+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 | 1.1.0          | 1.1.1:2 dev              |
 
 
+## 1.1.1 release
+### Added
+- `input()` and `output()` methods added to `droit.Database()`. They point to pythons `input()` and `print()` functions but can be changed. Plugins use those functions when needed. Therefore you can make sure every output goes to the device you want it to (e.g. when using text-to-speech).
+- get input-block added
+
+### Changed
+- output plugins have now the abillity to modify `inp` variables
 
 ## 1.1.1:2 dev
 ### Added